requirements.txt
apache-airflow==2.10.2
apache-airflow-providers-amazon==9.0.0
apache-airflow-providers-http
pyspark==3.5.3
delta-spark~=3.2.1